fix annotations

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -7,7 +7,7 @@ import typing
 import random
 
 
-async def async_generator() -> typing.Generator[int, None, None]:
+async def async_generator() -> typing.Generator[float, None, None]:
     """Generator function
 
     Yeilds:
